python/src/requirements.txt
colorama==0.4.6
pymongo==4.6.0
groq==0.31.0
ollama==0.1.6
openai==1.3.0
google-cloud-aiplatform==1.36.0
python-dotenv==1.0.0